some update

+++ src/FHEController.cpp
@@ -870,7 +870,7 @@ Ctxt FHEController::repeat(const Ctxt &in, int slots, int padding) {
 vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ptxt &weight, const Ptxt &bias) {
     vector<Ctxt> columns;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         Ctxt m = mult(rows[i], weight);
 
         m = rotsum(m, 128, 128);
@@ -886,7 +886,7 @@ vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ptxt &weight, cons
 vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ptxt &weight, const Ptxt &bias, int row_size, int padding) {
     vector<Ctxt> columns;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         Ctxt m = mult(rows[i], weight);
 
         m = rotsum(m, row_size, padding);
@@ -902,7 +902,7 @@ vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ptxt &weight, cons
 vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ctxt &weight, int row_size, int padding) {
     vector<Ctxt> columns;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         Ctxt m = mult(rows[i], weight);
 
         m = rotsum(m, row_size, padding);
@@ -916,7 +916,7 @@ vector<Ctxt> FHEController::matmulRE(vector<Ctxt> rows, const Ctxt &weight, int 
 vector<Ctxt> FHEController::matmulRElarge(vector<Ctxt>& inputs, const vector<Ptxt> &weights, const Ptxt &bias, double mask_val) {
     vector<Ctxt> densed;
 
-    for (int i = 0; i < inputs.size(); i++) {
+    for (int i = 0; i < (int)inputs.size(); i++) {
         Ctxt i_th_result;
         for (int j = weights.size() - 1; j >= 0; j--) {
             Ctxt out = mult(inputs[i], weights[j]);
@@ -924,7 +924,7 @@ vector<Ctxt> FHEController::matmulRElarge(vector<Ctxt>& inputs, const vector<Ptx
 
             out = mask_first_n(out, 128, mask_val);
 
-            if (j == weights.size() - 1)
+            if (j == (int)weights.size() - 1)
                 i_th_result = out;
             else {
                 //i_th_result = rotate(i_th_result, -128);
@@ -947,7 +947,7 @@ vector<Ctxt> FHEController::matmulRElarge(vector<Ctxt>& inputs, const vector<Ptx
 vector<Ctxt> FHEController::matmulCR(vector<Ctxt> rows, const Ctxt& matrix) {
     vector<Ctxt> columns;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         Ctxt m = mult(rows[i], matrix);
 
         m = rotsum(m, 64, 1);
@@ -961,7 +961,7 @@ vector<Ctxt> FHEController::matmulCR(vector<Ctxt> rows, const Ctxt& matrix) {
 vector<Ctxt> FHEController::matmulCR(vector<Ctxt> rows, const Ptxt& weight, const Ptxt& bias) {
     vector<Ctxt> columns;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         Ctxt m = mult(rows[i], weight);
 
         m = rotsum(m, 128, 1);
@@ -977,7 +977,7 @@ vector<Ctxt> FHEController::matmulCR(vector<Ctxt> rows, const Ptxt& weight, cons
 vector<Ctxt> FHEController::matmulCRlarge(vector<vector<Ctxt>> rows, vector<Ptxt> weights, const Ptxt &bias) {
     vector<Ctxt> output;
 
-    for (int i = 0; i < rows.size(); i++) {
+    for (int i = 0; i < (int)rows.size(); i++) {
         //Qua sotto posso fare prima add-many e poi un solo rotsum mi sa:)
         /*
         Ctxt p1 = rotsum(mult(rows[i][0], weights[0]), 128, 1);
@@ -1025,7 +1025,7 @@ Ctxt FHEController::matmulScores(vector<Ctxt> queries, const Ctxt &key) {
 Ctxt FHEController::wrapUpRepeated(vector<Ctxt> vectors) {
     vector<Ctxt> masked;
 
-    for (int i = 0; i < vectors.size(); i++) {
+    for (int i = 0; i < (int)vectors.size(); i++) {
         masked.push_back(mask_block(vectors[i], 128 * i, 128 * (i + 1), 1));
     }
 
@@ -1077,7 +1077,7 @@ vector<vector<Ctxt>> FHEController::unwrapRepeatedLarge(vector<Ctxt> containers,
         quantities.push_back(quantity);
     }
 
-    for (int i = 0; i < containers.size(); i++) {
+    for (int i = 0; i < (int)containers.size(); i++) {
         for (int j = 0; j < quantities[i]; j++) {
             vector<Ctxt> unwrapped_container = unwrap_512_in_4_128(containers[i], j);
             unwrapped_output.push_back(unwrapped_container);
@@ -1135,7 +1135,7 @@ vector<Ctxt> FHEController::generate_containers(vector<Ctxt> inputs, const Ptxt&
 
     for (int i = 0; i < inputs.size() / 32.0; i++) {
         int quantity = 32;
-        if ((i + 1) * 32 > inputs.size()) {
+        if ((i + 1) * 32 > (int)inputs.size()) {
             quantity = inputs.size() - (i * 32);
         }
 
@@ -1240,7 +1240,7 @@ Ctxt FHEController::eval_exp(const Ctxt &c, int inputs_number) {
     //Coefficients of Taylor series
     Ctxt res = context->EvalPoly(c, {1, 1, 1/(2.0), 1/(6.0), 1/(24.0), 1/(120.0), 1/(720.0)});
 
-    if (res->GetLevel() + 4 > circuit_depth) {
+    if ((int)(res->GetLevel() + 4) > circuit_depth) {
         res = bootstrap(res);
     }
 
@@ -1286,10 +1286,10 @@ Ctxt FHEController::eval_tanh_function(const Ctxt &c, double min, double max, do
 }
 
 vector<Ctxt> FHEController::slicing(vector<Ctxt> &arr, int X, int Y) {
-    if (Y - X >= arr.size())
+    if (Y - X >= (int)arr.size())
         return arr;
 
-    if (Y > arr.size()) {
+    if (Y > (int)arr.size()) {
         Y = arr.size();
     }
